Bump version
Also do not needlessly unload models in standalone mode.
