+++ scripts/depthmap.py
@@ -10,20 +10,6 @@ from src.gradio_args_transport import GradioComponentBundle
 from src.misc import *
 
 
-# Ugly workaround to fix gradio tempfile issue
-def ensure_gradio_temp_directory():
-    try:
-        import tempfile
-        path = os.path.join(tempfile.gettempdir(), 'gradio')
-        if not (os.path.exists(path)):
-            os.mkdir(path)
-    except Exception as e:
-        traceback.print_exc()
-
-
-ensure_gradio_temp_directory()
-
-
 class Script(scripts.Script):
     def title(self):
         return SCRIPT_NAME
